agents/search_agent/pyproject.toml
[project]
name = "search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai>=0.4.5",
    "langchain>=0.1.0",
    "langgraph>=0.1.0",
    "uvicorn>=0.24.0",
    "fastapi>=0.100.0",
    "httpx>=0.24.0",
    "langchain-openai>=0.1.0",
    "langchain-core>=0.1.0",
    "sseclient-py>=1.7.0",
    "a2a-sdk>=0.1.0",
    "langchain-mcp-adapters>=0.1.0",
]
